# Evaluation allocation improvements