--- a/.gitignore
+++ b/.gitignore
@@ -22,6 +22,9 @@ bld/
 [Bb]in/
 [Oo]bj/
 
+# Rider
+.idea/
+
 # Visual Studio 2015 cache/options directory
 .vs/