--- a/src/Directory.Packages.props
+++ b/src/Directory.Packages.props
@@ -17,5 +17,6 @@
     <PackageVersion Include="Spring.Core" Version="2.0.1" />
     <PackageVersion Include="System.CodeDom" Version="7.0.0" />
     <PackageVersion Include="System.Collections.Immutable" Version="7.0.0" />
+    <PackageVersion Include="System.Configuration.ConfigurationManager" Version="7.0.0" />
   </ItemGroup>
 </Project>