--- a/src/VCEL.Core/Expression/Impl/MaybeNotEqExpr.cs
+++ b/src/VCEL.Core/Expression/Impl/MaybeNotEqExpr.cs
@@ -5,7 +5,7 @@ using VCEL.Monad.Maybe;
 
 namespace VCEL.Core.Expression.Impl
 {
-    public class MaybeNotEqExpr<TMonad> : IExpression<TMonad> where TMonad : Maybe<object>
+    public class MaybeNotEqExpr<TMonad> : IExpression<TMonad> where TMonad : struct, IMaybe<object?>
     {
         public MaybeNotEqExpr(
             IMonad<TMonad> monad,