--- a/src/VCEL.Core/Expression/MissingFunctionException.cs
+++ b/src/VCEL.Core/Expression/MissingFunctionException.cs
@@ -17,6 +17,7 @@ namespace VCEL.Expression
         {
         }
 
+        [Obsolete("Obsolete")]
         protected MissingFunctionException(SerializationInfo info, StreamingContext context) : base(info, context)
         {
         }