… .github/workflows/dotnet-core-publish.yml
… .github/workflows/dotnet-core.yml
… src/Tests/Spel.Benchmark/LogicalEvaluatorBenchmarks.cs
… src/Tests/Spel.Benchmark/Program.cs
… src/Tests/Spel.Benchmark/Spel.Benchmark.csproj
… src/VCEL.Core/IContext.cs
… src/VCEL.Core/IValueAccessor.cs
… src/VCEL.Core/Monad/ConcatStringMonad.cs
… src/VCEL.Core/Monad/ExprMonad.cs
… src/VCEL.Core/Monad/IMonad.cs
… src/VCEL.Core/Monad/List/ListMonad.cs
… src/VCEL.Core/Monad/Maybe/Maybe.cs
… src/VCEL.Core/Monad/Maybe/MaybeMonad.cs
… src/VCEL.Core/Monad/Tasks/TaskMonad.cs
… src/VCEL.Core/ObjectContext.cs
… src/VCEL.Core/ObjectContextEx.cs
… src/VCEL.Core/OverrideAccessor.cs
… src/VCEL.Core/OverrideContext.cs
… src/VCEL.Core/PropertyValueAccessor.cs
… src/VCEL.JS/JsObjectContext.cs
… src/VCEL.JS/JsPropertyValueAccessor.cs
… src/VCEL.JS/ToJsCodeFactory.cs