## Main changes
- Remove closures where possible
- Move `IContext`/`IValueAccessor` to be structs, readonly where possible
- Update benchmarks

Evaluation allocations mainly now come from value type boxing and unboxing.

## Highlights
VcelDefaultAnd 248 B → 56 B
VcelMaybeAnd 280 B → 56 B
VcelMaybeTern 360 B → 136 B

## Before
| Method          | Mean      | Error      | StdDev   | Rank | Gen0   | Gen1   | Allocated |
|---------------- |----------:|-----------:|---------:|-----:|-------:|-------:|----------:|
| CSharpOr        |  13.02 ns |   8.247 ns | 0.452 ns |    1 | 0.0067 | 0.0000 |      56 B |
| CSharpAnd       |  13.04 ns |   0.774 ns | 0.042 ns |    1 | 0.0067 | 0.0000 |      56 B |
| CSharpNot       |  13.22 ns |   6.658 ns | 0.365 ns |    2 | 0.0067 |      - |      56 B |
| SpelNot         |  35.84 ns |  20.922 ns | 1.147 ns |    3 | 0.0086 |      - |      72 B |
| SpelOr          |  35.97 ns |  15.752 ns | 0.863 ns |    3 | 0.0086 |      - |      72 B |
| VcelDefaultOr   |  39.09 ns |  14.373 ns | 0.788 ns |    4 | 0.0220 | 0.0001 |     184 B |
| VcelMaybeOr     |  47.19 ns |  57.713 ns | 3.163 ns |    5 | 0.0258 |      - |     216 B |
| VcelDefaultAnd  |  51.43 ns |  14.840 ns | 0.813 ns |    6 | 0.0296 | 0.0001 |     248 B |
| SpelAnd         |  52.89 ns |   3.983 ns | 0.218 ns |    7 | 0.0086 |      - |      72 B |
| VcelMaybeAnd    |  55.35 ns |  15.051 ns | 0.825 ns |    8 | 0.0334 |      - |     280 B |
| VcelDefaultNot  |  66.09 ns |  49.006 ns | 2.686 ns |    9 | 0.0143 | 0.0001 |     120 B |
| VcelMaybeNot    |  74.07 ns |  92.847 ns | 5.089 ns |   10 | 0.0181 |      - |     152 B |
| CSharpTern      | 181.63 ns |  98.042 ns | 5.374 ns |   11 | 0.0324 |      - |     272 B |
| VcelDefaultTern | 207.18 ns |  79.348 ns | 4.349 ns |   12 | 0.0353 |      - |     296 B |
| VcelMaybeTern   | 216.48 ns |  96.338 ns | 5.281 ns |   13 | 0.0429 |      - |     360 B |
| SpelTern        | 227.58 ns | 135.860 ns | 7.447 ns |   14 | 0.0172 |      - |     144 B |

## After
| Method                       | Mean      | Error      | StdDev    | Gen0   | Gen1   | Gen2   | Allocated |
|----------------------------- |----------:|-----------:|----------:|-------:|-------:|-------:|----------:|
| SpelNotRow                   |  33.94 ns |  11.613 ns |  0.637 ns | 0.0086 |      - |      - |      72 B |
| SpelNotDynamicRow            |  35.21 ns |  37.163 ns |  2.037 ns | 0.0086 |      - |      - |      72 B |
| SpelNotDictionaryRow         |  34.20 ns |  21.833 ns |  1.197 ns | 0.0086 |      - |      - |      72 B |
| SpelAndRow                   |  52.29 ns |  11.253 ns |  0.617 ns | 0.0086 |      - |      - |      72 B |
| SpelAndDynamicRow            |  55.80 ns |  30.823 ns |  1.690 ns | 0.0086 |      - |      - |      72 B |
| SpelAndDictionaryRow         |  54.41 ns |  22.461 ns |  1.231 ns | 0.0086 |      - |      - |      72 B |
| SpelOrRow                    |  37.19 ns |  33.947 ns |  1.861 ns | 0.0086 |      - |      - |      72 B |
| SpelOrDynamicRow             |  36.89 ns |  11.888 ns |  0.652 ns | 0.0086 |      - |      - |      72 B |
| SpelOrDictionaryRow          |  37.71 ns |   5.825 ns |  0.319 ns | 0.0086 |      - |      - |      72 B |
| SpelTernRow                  | 242.38 ns |  45.655 ns |  2.503 ns | 0.0172 |      - |      - |     144 B |
| SpelTernDynamicRow           | 158.47 ns |  72.118 ns |  3.953 ns | 0.0114 |      - |      - |      96 B |
| VcelDefaultNotRow            |  78.91 ns |  22.579 ns |  1.238 ns | 0.0143 |      - |      - |     120 B |
| VcelDefaultNotDynamicRow     |  81.70 ns | 102.235 ns |  5.604 ns | 0.0143 |      - |      - |     120 B |
| VcelDefaultNotDictionaryRow  |  86.55 ns |  62.511 ns |  3.426 ns | 0.0143 |      - |      - |     120 B |
| VcelDefaultAndRow            |  44.50 ns |  35.181 ns |  1.928 ns | 0.0067 |      - |      - |      56 B |
| VcelDefaultAndDynamicRow     |  47.74 ns |  33.021 ns |  1.810 ns | 0.0067 |      - |      - |      56 B |
| VcelDefaultAndDictionaryRow  |  42.99 ns |  16.617 ns |  0.911 ns | 0.0067 |      - |      - |      56 B |
| VcelDefaultOrRow             |  32.31 ns |  37.681 ns |  2.065 ns | 0.0067 |      - |      - |      56 B |
| VcelDefaultOrDynamicRow      |  31.84 ns |  14.726 ns |  0.807 ns | 0.0067 |      - |      - |      56 B |
| VcelDefaultOrDictionaryRow   |  35.14 ns |  15.939 ns |  0.874 ns | 0.0067 |      - |      - |      56 B |
| VcelDefaultTernRow           | 200.44 ns |  31.247 ns |  1.713 ns | 0.0162 |      - |      - |     136 B |
| VcelDefaultTernDynamicRow    |  73.19 ns | 250.873 ns | 13.751 ns | 0.0038 |      - |      - |      32 B |
| VcelDefaultTernDictionaryRow |  65.65 ns |  61.825 ns |  3.389 ns | 0.0038 |      - |      - |      32 B |
| VcelMaybeNotRow              |  94.26 ns |  37.176 ns |  2.038 ns | 0.0143 |      - |      - |     120 B |
| VcelMaybeNotDynamicRow       |  85.58 ns |  20.437 ns |  1.120 ns | 0.0143 |      - |      - |     120 B |
| VcelMaybeNotDictionaryRow    | 108.78 ns |  19.052 ns |  1.044 ns | 0.0143 |      - |      - |     120 B |
| VcelMaybeAndRow              |  50.24 ns |  32.894 ns |  1.803 ns | 0.0067 |      - |      - |      56 B |
| VcelMaybeAndDynamicRow       |  49.92 ns |  12.783 ns |  0.701 ns | 0.0067 |      - |      - |      56 B |
| VcelMaybeAndDictionaryRow    |  48.50 ns |  49.096 ns |  2.691 ns | 0.0067 |      - |      - |      56 B |
| VcelMaybeOrRow               |  33.89 ns |  10.521 ns |  0.577 ns | 0.0067 |      - |      - |      56 B |
| VcelMaybeOrDynamicRow        |  34.24 ns |  14.022 ns |  0.769 ns | 0.0067 |      - |      - |      56 B |
| VcelMaybeOrDictionaryRow     |  34.30 ns |  11.221 ns |  0.615 ns | 0.0067 |      - |      - |      56 B |
| VcelMaybeTernRow             | 246.02 ns | 269.451 ns | 14.770 ns | 0.0162 |      - |      - |     136 B |
| VcelMaybeTernDynamicRow      |  80.52 ns |  38.873 ns |  2.131 ns | 0.0038 |      - |      - |      32 B |
| VcelMaybeTernDictionaryRow   |  77.29 ns |   7.575 ns |  0.415 ns | 0.0038 |      - |      - |      32 B |
| CSharpNotRow                 |  14.76 ns |   2.227 ns |  0.122 ns | 0.0067 |      - |      - |      56 B |
| CSharpNotDynamicRow          |  14.28 ns |   4.390 ns |  0.241 ns | 0.0067 |      - |      - |      56 B |
| CSharpNotDictionaryRow       |  14.56 ns |  12.488 ns |  0.684 ns | 0.0067 |      - |      - |      56 B |
| CSharpAndRow                 |  15.44 ns |  10.132 ns |  0.555 ns | 0.0067 |      - |      - |      56 B |
| CSharpAndDynamicRow          |  15.28 ns |  11.920 ns |  0.653 ns | 0.0067 |      - |      - |      56 B |
| CSharpAndDictionaryRow       |  14.77 ns |   1.908 ns |  0.105 ns | 0.0067 |      - |      - |      56 B |
| CSharpOrRow                  |  14.13 ns |   5.193 ns |  0.285 ns | 0.0067 | 0.0000 | 0.0000 |      56 B |
| CSharpOrDynamicRow           |  14.67 ns |   5.921 ns |  0.325 ns | 0.0067 | 0.0000 |      - |      56 B |
| CSharpOrDictionaryRow        |  14.81 ns |  10.452 ns |  0.573 ns | 0.0067 |      - |      - |      56 B |
| CSharpTernRow                | 179.54 ns | 111.751 ns |  6.125 ns | 0.0324 | 0.0002 |      - |     272 B |
| CSharpTernDynamicRow         |  92.57 ns |  35.459 ns |  1.944 ns | 0.0267 |      - |      - |     224 B |